! Fixed another bug

+++ src/bugs/bugplayer.cpp
@@ -110,7 +110,7 @@ void BugPlayer::DrawFloor() {
 void BugPlayer::DrawTop() {
   CL_Colorf color = CL_Colorf::white;
     default_font.draw_text(*gc_, CL_Pointf(10, 30),
-                           cl_format("Bugs: %1", flies.size()), color);
+                           cl_format("Bugs: %1", (int)flies.size()), color);
   Player::DrawTop();
 }
